% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createpixid.R
\name{createPixid}
\alias{createPixid}
\title{Create pixel id grid}
\usage{
createPixid(xmin = 0, xmax = 280000, ymin = 3e+05, ymax = 625000,
  resolution = 25)
}
\arguments{
\item{xmin}{minimum x coordinate}

\item{xmax}{maximum x coordinate}

\item{ymin}{minimum y coordinate}

\item{ymax}{maximum y coordinate}

\item{resolution}{resolution of grid}
}
\value{
A raster grid with extent (xmin,xmax,ymin,ymax) and given
resolution. All values in meters. The 'Amersfoort' projection is
used as coordinate reference system.
}
\description{
Creates a pixel id grid, a grid with consecutive cell numbers.
}
\details{
This functions requires a datacube reference (pixid) grid
}
