% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcget.R
\name{dcget}
\alias{dcget}
\title{Get blob object from datacube}
\usage{
dcget(obj, update = FALSE, verbose = FALSE)
}
\arguments{
\item{obj}{name of blob object}

\item{update}{force update}

\item{verbose}{show extra verbose output}
}
\value{
This function returns the pgobjects blob object. If the object is
updated, i.e an updated file is copied from the blobpath, than an
audit object is created which records the data lineage. The
returned updated object includes this audit object.
}
\description{
Gets a blob object from the datacube. This function checks if the
blob already exists and, if so, needs to be updated. Also this
function creates an audit object for recording the data lineage
}
