% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcimport.R
\name{dcimport}
\alias{dcimport}
\title{Import datacube dataset from external file path}
\usage{
dcimport(importfile, overwrite = FALSE)
}
\arguments{
\item{importfile}{full path of the file to import}

\item{overwrite}{Should existing blobs be overwritten

Please note that the importBLob function overwrites existing files in the
data directory}
}
\value{
a pgobjects blob object
}
\description{
This function imports a datacube data set, i.e. a single file with
data and the ini file with meta data, from a location on disk. This
function is a wrapper around the pgblobs::blobImport function.
}
