% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcstore.R
\name{dcstore}
\alias{dcstore}
\title{Stores a single file as blob object in the datacube}
\usage{
dcstore(filename, objname, kv)
}
\arguments{
\item{filename}{filename of file to store}

\item{objname}{name of the object to store}

\item{kv}{list of key, pairs}
}
\value{
This function returns the pgobjects blob object, which includes the
object meta data
}
\description{
This function stores a single file as blob object in the datacube.
Meta data is given as a list with key,value pairs
}
