% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacube.R
\docType{data}
\name{.DatacubeConfig}
\alias{.DatacubeConfig}
\title{Get git info}
\format{An object of class \code{environment} of length 0.}
\usage{
.DatacubeConfig
}
\description{
Get information about git repository
}
\keyword{datasets}
