% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getobjlist.R
\name{getobjlist}
\alias{getobjlist}
\title{Show datacube objects in database}
\usage{
getobjlist(schema = "datacube")
}
\arguments{
\item{schema}{database schema where the objects are stored}
}
\value{
data.frame with object names, description and automatically
generated meta data
}
\description{
This function returns a data.frame with the name and description of all the objects and their
automatically generated metadata
}
