% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getprojcfg.R
\name{getprojcfg}
\alias{getprojcfg}
\title{get project config}
\usage{
getprojcfg(dir, script, setdir = TRUE)
}
\arguments{
\item{dir}{project directory}

\item{script}{filename of R script}

\item{setdir}{if TRUE, change workingdirectory to project dir}
}
\description{
get project config
}
