% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getraster.R
\name{getRaster}
\alias{getRaster}
\title{Get a raster from the datacube}
\usage{
getRaster(obj, path = "./data/", update = FALSE)
}
\arguments{
\item{obj}{name of datacube object}

\item{path}{path to store grid file}

\item{update}{if TRUE, always get raster even if it allready exists}
}
\value{
the meta data of the blob object

This a wrapper function around \code{link{pgblobs::getBlob}}.  
It assumes that rasters in the datacube are stored as GeoTIFF
files. This function gets the GeoTiff file from the datacube and
converts it into a normal raster file. This function also assumes
that the object name is the same as the basename of the raster
file without extention.

This functions checks if the raster file allready exists, if it
exists is does not load the file from the datacube since this is
time consuming. If the user wants to update the file, use argument
`update=TRUE`

The GeoTiff will be deleted after writing the raster file.
}
\description{
This function loads a raster grid file from the datacube and stores
it into the data directory
}
