% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid2tiff.R
\name{grid2tif}
\alias{grid2tif}
\title{converts a raster grid to GeoTIFF}
\usage{
grid2tif(gridfile)
}
\arguments{
\item{gridfile}{filename of the raster grid file}
}
\value{
the file path of the GeoTIFF file
}
\description{
This function converts a raster grid file (actually, 2 files) into
a GeoTIFF file (a single file). The GeoTIFF file get's the same
file name as the raster file. If the GeoTIFF file exists, it will
be overwriten.
}
