% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_postgis.R
\name{read_postgis}
\alias{read_postgis}
\title{Read a spatial vector object from postgis}
\usage{
read_postgis(layername)
}
\arguments{
\item{layername}{name of layer}
}
\value{
returns sf geo object
}
\description{
Write a spatial vector object to the datacube postgis tables
}
