% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storekvlist.R
\name{storekvlist}
\alias{storekvlist}
\title{store key-value pairs}
\usage{
storekvlist(kv, name)
}
\description{
Key-value pairs can be stored in pgobjects database, this function
stores key-value pairs based on a list
}
