% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid2blob.R
\name{storeRaster}
\alias{storeRaster}
\title{Store a raster into the datacube}
\usage{
storeRaster(gridfile, blobname, kv, desc)
}
\arguments{
\item{gridfile}{filename of the raster grid file}

\item{blobname}{objectname of the blob}

\item{kv}{meta data key-value pairs}

\item{desc}{blob object description}
}
\value{
the meta data of the blob object

This a wrapper function around \code{link{grid2tif}} and
\code{link{pgblobs::createBlob}}. It first converts the raster file
into a GeoTIFF file and then stores the GeoTIFF file into the
datacube. The GeoTIFF file is not removed from the data
directory

please not that this function results in an error if the blob
object allready exists.
}
\description{
This function stores a raster grid file as blob into the datacube
}
