% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_posgis.R
\name{write_postgis}
\alias{write_postgis}
\title{Write a spatial vector object to postgis}
\usage{
write_postgis(obj, layername, schema = "gis")
}
\arguments{
\item{obj}{spatial vector object}

\item{layername}{name of layer}

\item{schema}{database schema were to create the table}
}
\value{
does not return anything
}
\description{
Write a spatial vector object to the datacube postgis tables
}
