% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacubeInit.R
\name{datacubeInit}
\alias{datacubeInit}
\title{Init Datacube}
\usage{
datacubeInit(script, workdir, database = "datacube",
  optionsfile = "~/.R.options", setdir = TRUE)
}
\description{
initialiaze datacube
}
