% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacube.R
\docType{data}
\name{.DatacubeConfig}
\alias{.DatacubeConfig}
\title{Setup of datacube configuration environment}
\format{An object of class \code{environment} of length 0.}
\usage{
.DatacubeConfig
}
\description{
Setup of datacube configuration environment
}
\keyword{datasets}
