% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_postgis.R
\name{read_postgis}
\alias{read_postgis}
\title{Read a spatial vector object from postgis}
\usage{
read_postgis(layername)
}
\arguments{
\item{layername}{name of layer}
}
\value{
a sf object with the postgis layer
}
\description{
Read a spatial vector object from the datacube postgis tables
}
